% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\docType{data}
\name{nodes}
\alias{nodes}
\title{Data on documents in the acquis communautaire}
\format{
A data frame with 9 variables:
\describe{
\item{key_id}{\code{numeric}. An ID number that uniquely identifies each
observation. Indicates the default sorting order.}
\item{celex}{\code{string}. The CELEX number for the document.}
\item{node_type_id}{\code{numeric}. An ID number that uniquely identifies
the type of the outgoing node.}
\item{node_type}{\code{string}. The type of the node. Possible values
include: \code{Treaty articles}, \code{Directives}, \code{Regulations}.
\code{Decisions}, \code{Court of Justice judgments}, \code{Advocate general
opinions}, \code{General Court judgments}, \code{National implementing
measures}, and \code{National court decisions}.}
\item{author}{\code{string}. The author of the document. If there are
multiple authors, they are listed, separated by a comma.}
\item{date}{\code{date}. The date of the document in a \code{YYYY-MM-DD}
format. Coded \code{NA} for national implementing measures.}
\item{year}{\code{numeric}. The year of the document. Coded \code{NA} for
national implementing measures.}
\item{month}{\code{numeric}. The month of the document. Coded \code{NA} for
national implementing measures.}
\item{day}{\code{numeric}. The day of the document. Coded \code{NA} for
national implementing measures.}
}
}
\usage{
nodes
}
\description{
This dataset includes one one observation per node in the network of
European Union legal documents. Each node is a document. This network
includes primary law (i.e., treaty articles), secondary law (i.e.,
legislation), and case law (i.e., court cases). The dataset indicates the
CELEX number of each document, which is a unique ID number assigned by
EUR-Lex, the date of the document, the author of the document, and the type
of the document. The dataset includes the following types of documents:
treaty articles, directives, regulations, decisions, Court of Justice
judgments, advocate general opinions, General Court judgments, national
implementing measures (i.e., national legislation that member states enact
to transpose EU directives), and national court decisions that apply EU
law.
}
\keyword{datasets}
